Add missing onPreCompile stage hook
