+++ test/sources/projects/task-hooks/contracts/Hooks.sol
@@ -1,0 +1,17 @@
+pragma solidity >=0.8.0 <0.9.0;
+
+
+contract ContractA {
+  uint x;
+  constructor() public {
+  }
+
+  function sendFn() public {
+    x = 5;
+  }
+
+  function callFn() public pure returns (uint){
+    uint y = 5;
+    return y;
+  }
+}
