+++ test/sources/projects/task-hooks/test/hooks.js
@@ -1,0 +1,15 @@
+const ContractA = artifacts.require("ContractA");
+
+contract("contracta", function(accounts) {
+  let instance;
+
+  before(async () => instance = await ContractA.new())
+
+  it('sends', async function(){
+    await instance.sendFn();
+  });
+
+  it('calls', async function(){
+    await instance.callFn();
+  })
+});
