+++ test/sources/projects/test-files/.solcover.js
@@ -1,13 +1,5 @@
-// Testing hooks
-const fn = (msg, config) => config.logger.log(msg);
-
 module.exports = {
-  skipFiles: ['Migrations.sol'],
+  skipFiles: [],
   silent: process.env.SILENT ? true : false,
-  istanbulReporter: ['json-summary', 'text'],
-  onPreCompile: fn.bind(null, 'running onPreCompile'),
-  onServerReady: fn.bind(null, 'running onServerReady'),
-  onTestsComplete: fn.bind(null, 'running onTestsComplete'),
-  onCompileComplete: fn.bind(null, 'running onCompileComplete'),
-  onIstanbulComplete: fn.bind(null, 'running onIstanbulComplete')
+  istanbulReporter: ['json-summary', 'text']
 }
